Add ScoreboardNicknameUtils and outsource utility methods from main service
